fix(telemetry+billing): address code-review findings
- #1 auto-top-up RATE guard: cap successful auto-top-ups per rolling 24h
  (migration 0022); trips the existing breaker so a succeeding runaway loop
  can't keep charging the card. The failure breaker only caught failed charges.
- #2 capture AI cost at the BearerClient chokepoint: post_json("/v1/ai/*") now
  records app_ai_calls, so compaction / day_summary / day_illustration /
  narrative_identity (and any future non-streaming AI feature) are no longer
  invisible. Background callers tag .with_feature(...); transcription's explicit
  capture removed (now chokepoint-captured).
- #7 chat records its real feature (agent_mode: chat | council | deep_research)
  instead of a hardcoded "chat".
- #8 thread reasoning_tokens through AgentEvent::Usage into app_ai_calls.
- #6 drop the duplicate usd_to_micros in routes/ai.rs (use entitlement's).
- #4 app_system_samples.load1 → DOUBLE PRECISION (matches the f64 source;
  no more silent f64→f32 narrowing); SystemSample.load1 → Option<f64>.

+++ virtues-core/src/api/system_telemetry.rs
@@ -836,7 +836,7 @@ pub struct SystemSample {
     pub disk_used_bytes: Option<i64>,
     pub disk_total_bytes: Option<i64>,
     pub temp_c: Option<f32>,
-    pub load1: Option<f32>,
+    pub load1: Option<f64>,
     pub sidecar_embed_up: Option<bool>,
     pub sidecar_rerank_up: Option<bool>,
 }
